Fixed Issue #359
